\name{PretendMachinist-package}
\alias{PretendMachinist-package}
\alias{PretendMachinist}
\docType{package}
\title{
\packageTitle{PretendMachinist}
}
\description{
\packageDescription{PretendMachinist}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PretendMachinist}
\packageIndices{PretendMachinist}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PretendMachinist}

Maintainer: \packageMaintainer{PretendMachinist}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
